requirements.txt
streamlit==1.15.1
redmail
PILLOW
